[DEBO10C01][DuongHT] update landing page
