+++ src/components/Appointment/AppointmentDetail/AppointmentNotes.tsx
@@ -1,276 +1,29 @@
-import { useState, useEffect, useRef } from 'react';
-import { CKEditor } from '@ckeditor/ckeditor5-react';
+import MyCKEditor from "@/components/CKEditor/MyCKEditor";
+import { toastError, toastSuccess } from "@/utils/toast";
 
-import {
-  ClassicEditor,
-  AccessibilityHelp,
-  Autoformat,
-  AutoImage,
-  AutoLink,
-  Autosave,
-  BalloonToolbar,
-  Base64UploadAdapter,
-  BlockQuote,
-  Bold,
-  Code,
-  CodeBlock,
-  Essentials,
-  FindAndReplace,
-  Heading,
-  Highlight,
-  HorizontalLine,
-  HtmlEmbed,
-  ImageBlock,
-  ImageCaption,
-  ImageInline,
-  ImageInsert,
-  ImageInsertViaUrl,
-  ImageResize,
-  ImageStyle,
-  ImageTextAlternative,
-  ImageToolbar,
-  ImageUpload,
-  Indent,
-  IndentBlock,
-  Italic,
-  Link,
-  LinkImage,
-  List,
-  ListProperties,
-  Markdown,
-  MediaEmbed,
-  Mention,
-  Paragraph,
-  PasteFromMarkdownExperimental,
-  PasteFromOffice,
-  SelectAll,
-  SpecialCharacters,
-  SpecialCharactersArrows,
-  SpecialCharactersCurrency,
-  SpecialCharactersEssentials,
-  SpecialCharactersLatin,
-  SpecialCharactersMathematical,
-  SpecialCharactersText,
-  Strikethrough,
-  Table,
-  TableCellProperties,
-  TableProperties,
-  TableToolbar,
-  TextTransformation,
-  TodoList,
-  Underline,
-  Undo
-} from 'ckeditor5';
+function AppointmentNotes(url: any) {
+  const handleEditorSubmit = (editorData: any) => {
 
-import 'ckeditor5/ckeditor5.css';
-
-import './Styles/style.css';
-
-export default function AppointmentNotes() {
-  const editorContainerRef = useRef(null);
-  const editorRef = useRef(null);
-  const [isLayoutReady, setIsLayoutReady] = useState(false);
-
-  useEffect(() => {
-    setIsLayoutReady(true);
-
-    return () => setIsLayoutReady(false);
-  }, []);
-
-  const editorConfig = {
-    toolbar: {
-      items: [
-        'undo',
-        'redo',
-        '|',
-        'heading',
-        '|',
-        'bold',
-        'italic',
-        'underline',
-        '|',
-        'link',
-        'insertImage',
-        'insertTable',
-        'highlight',
-        'blockQuote',
-        'codeBlock',
-        '|',
-        'bulletedList',
-        'numberedList',
-        'todoList',
-        'indent',
-        'outdent'
-      ],
-      shouldNotGroupWhenFull: false
-    },
-    plugins: [
-      AccessibilityHelp,
-      Autoformat,
-      AutoImage,
-      AutoLink,
-      Autosave,
-      BalloonToolbar,
-      Base64UploadAdapter,
-      BlockQuote,
-      Bold,
-      Code,
-      CodeBlock,
-      Essentials,
-      FindAndReplace,
-      Heading,
-      Highlight,
-      HorizontalLine,
-      HtmlEmbed,
-      ImageBlock,
-      ImageCaption,
-      ImageInline,
-      ImageInsert,
-      ImageInsertViaUrl,
-      ImageResize,
-      ImageStyle,
-      ImageTextAlternative,
-      ImageToolbar,
-      ImageUpload,
-      Indent,
-      IndentBlock,
-      Italic,
-      Link,
-      LinkImage,
-      List,
-      ListProperties,
-      Markdown,
-      MediaEmbed,
-      Mention,
-      Paragraph,
-      PasteFromMarkdownExperimental,
-      PasteFromOffice,
-      SelectAll,
-      SpecialCharacters,
-      SpecialCharactersArrows,
-      SpecialCharactersCurrency,
-      SpecialCharactersEssentials,
-      SpecialCharactersLatin,
-      SpecialCharactersMathematical,
-      SpecialCharactersText,
-      Strikethrough,
-      Table,
-      TableCellProperties,
-      TableProperties,
-      TableToolbar,
-      TextTransformation,
-      TodoList,
-      Underline,
-      Undo
-    ],
-    balloonToolbar: ['bold', 'italic', '|', 'link', 'insertImage', '|', 'bulletedList', 'numberedList'],
-    heading: {
-      options: [
-        {
-          model: 'paragraph',
-          title: 'Paragraph',
-          class: 'ck-heading_paragraph'
-        },
-        {
-          model: 'heading1',
-          view: 'h1',
-          title: 'Heading 1',
-          class: 'ck-heading_heading1'
-        },
-        {
-          model: 'heading2',
-          view: 'h2',
-          title: 'Heading 2',
-          class: 'ck-heading_heading2'
-        },
-        {
-          model: 'heading3',
-          view: 'h3',
-          title: 'Heading 3',
-          class: 'ck-heading_heading3'
-        },
-        {
-          model: 'heading4',
-          view: 'h4',
-          title: 'Heading 4',
-          class: 'ck-heading_heading4'
-        },
-        {
-          model: 'heading5',
-          view: 'h5',
-          title: 'Heading 5',
-          class: 'ck-heading_heading5'
-        },
-        {
-          model: 'heading6',
-          view: 'h6',
-          title: 'Heading 6',
-          class: 'ck-heading_heading6'
-        }
-      ]
-    },
-    image: {
-      toolbar: [
-        'toggleImageCaption',
-        'imageTextAlternative',
-        '|',
-        'imageStyle:inline',
-        'imageStyle:wrapText',
-        'imageStyle:breakText',
-        '|',
-        'resizeImage'
-      ]
-    },
-    initialData:
-      'Hello Doctor',
-    link: {
-      addTargetToExternalLinks: true,
-      defaultProtocol: 'https://',
-      decorators: {
-        toggleDownloadable: {
-          mode: 'manual',
-          label: 'Downloadable',
-          attributes: {
-            download: 'file'
-          }
-        }
-      }
-    },
-    list: {
-      properties: {
-        styles: true,
-        startIndex: true,
-        reversed: true
-      }
-    },
-    mention: {
-      feeds: [
-        {
-          marker: '@',
-          feed: [
-            /* See: https://ckeditor.com/docs/ckeditor5/latest/features/mentions.html */
-          ]
-        }
-      ]
-    },
-    menuBar: {
-      isVisible: true
-    },
-    placeholder: 'Type or paste your content here!',
-    table: {
-      contentToolbar: ['tableColumn', 'tableRow', 'mergeTableCells', 'tableProperties', 'tableCellProperties']
-    }
+    fetch(url, {
+      method: 'POST',
+      headers: {
+        'Content-Type': 'application/json',
+      },
+      body: JSON.stringify({ content: editorData }),
+    })
+      .then(response => response.json())
+      .then(data => {
+        console.log('Success:', data);
+        toastSuccess('Note saved successfully');
+      })
+      .catch((error) => {
+        toastError('Failed to save note');
+        console.error('Error:', error);
+      });
   };
-
   return (
-    <div>
-      <div className="main-container">
-        <div className="editor-container editor-container_classic-editor" ref={editorContainerRef}>
-          <div className="editor-container__editor">
-            <div ref={editorRef}>{isLayoutReady && <CKEditor editor={ClassicEditor} config={editorConfig} />}</div>
-          </div>
-        </div>
-      </div>
-    </div>
+    <MyCKEditor onSubmit={handleEditorSubmit} />
   );
 }
+
+export default AppointmentNotes;
