+++ src/components/CKEditor/Styles/style.css
@@ -20,7 +20,5 @@
 }
 
 .editor-container_classic-editor .editor-container__editor {
-    min-width: 795px;
-    max-width: 795px;
-
+    width: 100%;
 }
