+++ src/pages/Dentist/Appointment/DentistAppointmentNotes.tsx
@@ -1,0 +1,9 @@
+import AppointmentDetail from "@/components/Appointment/AppointmentDetail";
+
+function DentistAppointmentNotes() {
+    return (
+        <AppointmentDetail url={'My url'} />
+    );
+}
+
+export default DentistAppointmentNotes;
